Move MonitorRestEndpoint to testware
This should make it available outside core DC.

Fixes #1607
