+++ monitor/ui/pom.xml
@@ -423,7 +423,6 @@
 		<dependency>
 			<groupId>io.rest-assured</groupId>
 			<artifactId>rest-assured</artifactId>
-			<version>3.0.1</version>
 			<scope>test</scope>
 		</dependency>
 	</dependencies>
