+++ monitor/ui/src/test/java/org/datacleaner/monitor/JobServicesIT.java
@@ -23,6 +23,7 @@ import java.io.IOException;
 import java.net.URISyntaxException;
 
 import org.apache.http.HttpStatus;
+import org.datacleaner.test.MonitorRestEndpoint;
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
