+++ pom.xml
@@ -1452,6 +1452,11 @@
 				<scope>test</scope>
 			</dependency>
 			<dependency>
+				<groupId>io.rest-assured</groupId>
+				<artifactId>rest-assured</artifactId>
+				<version>3.0.1</version>
+			</dependency>
+			<dependency>
 				<groupId>org.scalatest</groupId>
 				<artifactId>scalatest_2.11</artifactId>
 				<version>2.2.0</version>
