+++ testware/src/main/java/org/datacleaner/test/MonitorRestEndpoint.java
@@ -17,7 +17,7 @@
  * 51 Franklin Street, Fifth Floor
  * Boston, MA  02110-1301  USA
  */
-package org.datacleaner.monitor;
+package org.datacleaner.test;
 
 import java.net.URI;
 import java.net.URISyntaxException;
